refactor: 복잡한 로직 훅으로 분리
[#802]
